[JBIDE-25700] moved OCBinary and validator to main package
