+++ plugins/org.jboss.tools.openshift.core/src/org/jboss/tools/openshift/internal/core/OCBinary.java
@@ -8,7 +8,7 @@
  * Contributors: 
  * Red Hat, Inc. - initial API and implementation 
  ******************************************************************************/
-package org.jboss.tools.openshift.internal.core.preferences;
+package org.jboss.tools.openshift.internal.core;
 
 import org.eclipse.core.runtime.Platform;
 import org.jboss.tools.openshift.common.core.connection.IConnection;
