+++ plugins/org.jboss.tools.openshift.core/src/org/jboss/tools/openshift/internal/core/OCBinaryValidator.java
@@ -8,7 +8,7 @@
  * Contributors: 
  * Red Hat, Inc. - initial API and implementation 
  ******************************************************************************/
-package org.jboss.tools.openshift.internal.core.preferences;
+package org.jboss.tools.openshift.internal.core;
 
 import java.io.BufferedReader;
 import java.io.File;
@@ -29,7 +29,6 @@ import org.eclipse.core.runtime.SubMonitor;
 import org.eclipse.osgi.util.NLS;
 import org.jboss.tools.foundation.core.plugin.log.StatusFactory;
 import org.jboss.tools.openshift.core.OpenShiftCoreMessages;
-import org.jboss.tools.openshift.internal.core.OpenShiftCoreActivator;
 import org.osgi.framework.Version;
 
 public class OCBinaryValidator {
